+++ picasso/gui/spinna.py
@@ -328,9 +328,13 @@ class MaskPreview(QtWidgets.QLabel):
     def save_current_view(self) -> None:
         """Save self.image (QImage, the current view) as png or tif."""
         path, _ = QtWidgets.QFileDialog.getSaveFileName(
-            self, "Save current view", filter="*.png;;*.tif"
+            self,
+            "Save current view",
+            directory=self.mask_tab.window.pwd,
+            filter="*.png;;*.tif",
         )
         if path:
+            self.mask_tab.window.pwd = os.path.dirname(path)
             self.qimage.save(path)
 
     def viewport_size(self) -> tuple[int, int] | None:
@@ -502,6 +506,7 @@ class MaskGeneratorTab(QtWidgets.QDialog):
         layout = QtWidgets.QGridLayout(self)
         self.setLayout(layout)
         self.preview = MaskPreview(self)
+        self.window = window
 
         self.locs_path = ""
         self.locs = None
@@ -687,9 +692,13 @@ class MaskGeneratorTab(QtWidgets.QDialog):
         """Load localizations / molecules for mask generation."""
         # get localizations file
         self.locs_path, _ = QtWidgets.QFileDialog.getOpenFileName(
-            self, "Load molecules for mask generation", filter="*.hdf5"
+            self,
+            "Load molecules for mask generation",
+            directory=self.window.pwd,
+            filter="*.hdf5",
         )
         if self.locs_path:
+            self.window.pwd = os.path.dirname(self.locs_path)
             self.mask_generator = spinna.MaskGenerator(self.locs_path)
             self.mask_ndim.clear()
             if hasattr(self.mask_generator.locs, "z"):
@@ -1273,6 +1282,7 @@ class StructuresTab(QtWidgets.QDialog):
         self.setAutoFillBackground(True)
         layout = QtWidgets.QGridLayout(self)
         self.setLayout(layout)
+        self.window = window
 
         self.structures = []
         self.current_structure = None
@@ -1467,9 +1477,13 @@ class StructuresTab(QtWidgets.QDialog):
         """Save current structures as a .yaml file."""
         self.update_current_structure()  # in case it was not saved yet
         path, _ = QtWidgets.QFileDialog.getSaveFileName(
-            self, "Save structures", filter="*.yaml"
+            self,
+            "Save structures",
+            directory=self.window.pwd,
+            filter="*.yaml",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             info = []
             for structure in self.structures:
                 m_info = {
@@ -1486,9 +1500,13 @@ class StructuresTab(QtWidgets.QDialog):
     def load_structures(self) -> None:
         """Load structures from in a .yaml file."""
         path, _ = QtWidgets.QFileDialog.getOpenFileName(
-            self, "Load structures", filter="*.yaml"
+            self,
+            "Load structures",
+            directory=self.window.pwd,
+            filter="*.yaml",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             with open(path, 'r') as file:
                 try:
                     info = list(yaml.load_all(file, Loader=yaml.FullLoader))
@@ -1666,9 +1684,13 @@ class StructuresTab(QtWidgets.QDialog):
     def save_preview(self) -> None:
         """Save current preview."""
         path, _ = QtWidgets.QFileDialog.getSaveFileName(
-            self, "Save current view", filter="*.png;;.tif"
+            self,
+            "Save current view",
+            directory=self.window.pwd,
+            filter="*.png;;*.tif",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             self.preview.qimage.save(path)
 
 
@@ -1728,7 +1750,7 @@ class GenerateSearchSpaceDialog(QtWidgets.QDialog):
         self.buttons.rejected.connect(self.reject)
 
     @staticmethod
-    def getParams(parent: QtWidgets.QWidget) -> list[int | bool]:
+    def getParams(parent: QtWidgets.QWidget) -> list:
         """Create the dialog and returns the numbers of molecular
         targets per simulation, number of simulations, resolution
         factor, check if the results are to be saved."""
@@ -1899,11 +1921,14 @@ class CompareModelsDialog(QtWidgets.QDialog):
     def on_add_model(self) -> None:
         """Add a new model (list of structures) to the dialog."""
         paths, ext = QtWidgets.QFileDialog.getOpenFileNames(
-            self, "Choose model(s)", filter="*.yaml"
+            self,
+            "Choose model(s)",
+            directory=self.sim_tab.window.pwd,
+            filter="*.yaml"
         )
         if not paths:
             return
-
+        self.sim_tab.window.pwd = os.path.dirname(paths[0])
         for path in paths:
             structures, targets = spinna.load_structures(path)
             # check that the loaded targets match the exp. data's targets
@@ -2339,7 +2364,7 @@ class SimulationsTab(QtWidgets.QDialog):
         only).
     depth_stack : QtWidgets.QStackedWidget
         Stack of widgets for setting the depth of the simulation.
-        Index == 0 ->
+        Index == 0 -> disabled (2D), index == 1 -> enabled (3D).
     dim_widget : QtWidgets.QComboBox
         Combo box for setting the dimension of the simulation (2D/3D).
     exp_data : dict
@@ -2441,6 +2466,7 @@ class SimulationsTab(QtWidgets.QDialog):
 
     def __init__(self, window: QtWidgets.QMainWindow) -> None:
         super().__init__(window)
+        self.window = window
         self.setAutoFillBackground(True)
         layout = QtWidgets.QGridLayout(self)
         left_column = QtWidgets.QGridLayout()
@@ -2674,9 +2700,13 @@ class SimulationsTab(QtWidgets.QDialog):
     def load_structures(self) -> None:
         """Load structures from .yaml file."""
         path, _ = QtWidgets.QFileDialog.getOpenFileName(
-            self, "Load structures", filter="*.yaml"
+            self,
+            "Load structures",
+            directory=self.window.pwd,
+            filter="*.yaml",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             self.structures, self.targets = spinna.load_structures(path)
 
             self.structures_path = path
@@ -2723,9 +2753,13 @@ class SimulationsTab(QtWidgets.QDialog):
         target species."""
         target = name[3:]
         path, _ = QtWidgets.QFileDialog.getOpenFileName(
-            self, f"Load molecules for {target}", filter="*.hdf5"
+            self,
+            f"Load molecules for {target}",
+            directory=self.window.pwd,
+            filter="*.hdf5",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             # load the data
             locs, info = io.load_locs(path)
             pixelsize = None
@@ -2768,9 +2802,13 @@ class SimulationsTab(QtWidgets.QDialog):
         """Load mask for the given molecular target species."""
         target = name[4:]
         path, _ = QtWidgets.QFileDialog.getOpenFileName(
-            self, f"Load mask for {target}", filter="*.npy"
+            self,
+            f"Load mask for {target}",
+            directory=self.window.pwd,
+            filter="*.npy",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             # load the data
             mask, info = io.load_mask(path)
             self.masks[target] = mask
@@ -2929,6 +2967,9 @@ class SimulationsTab(QtWidgets.QDialog):
             title_spin.setDecimals(2)
             title_spin.setSingleStep(1)
             title_spin.setValue(0)
+            title_spin.valueChanged.connect(
+                partial(self.check_prop_str_sum, name=title)
+            )
 
             label = QtWidgets.QLabel(f"{title}:")
             column = 2 if len(self.prop_str_input_spins) % 2 else 0
@@ -2979,6 +3020,7 @@ class SimulationsTab(QtWidgets.QDialog):
             )
             if not save:
                 return
+            self.window.pwd = os.path.dirname(save)
         self.n_sim_fit = n_sim_fit
         self.granularity = granularity
         self.n_sim_plot_spin.setValue(n_sim_fit)  # save in  NND plot settings
@@ -3038,7 +3080,7 @@ class SimulationsTab(QtWidgets.QDialog):
             gt_coords=self.exp_data,
             N_sim=self.n_sim_fit,
         )
-        _ = spinner.NN_scorer(N_structures, callback=None)
+        _ = spinner.NN_scorer(N_structures)
         dt = time.time() - t0
 
         # estimate the time for n combinations with a certain number of CPUs;
@@ -3058,21 +3100,42 @@ class SimulationsTab(QtWidgets.QDialog):
     def load_search_space(self) -> None:
         """Load combinations of numbers of structures for fitting."""
         path, _ = QtWidgets.QFileDialog.getOpenFileName(
-            self, "Load numbers of structures", filter="*.csv"
+            self,
+            "Load numbers of structures",
+            directory=self.window.pwd,
+            filter="*.csv",
         )
         if path:
+            self.window.pwd = os.path.dirname(path)
             df = pd.read_csv(path)
             # check if the structure titles are the same as the ones loaded
             loaded = df.columns
             titles = [_.title for _ in self.structures]
             if (
-                len(loaded) == len(titles)
-                and all([t in loaded for t in titles])
-            ):
+                len(loaded) == len(titles) * 2
+                and all([f"N_{t}" in loaded for t in titles])
+            ):  # check that all titles are present
                 self.N_structures_fit = {
-                    column: df[column].values.astype(np.int32)
-                    for column in df.columns
+                    structure_name: (
+                        df[f"N_{structure_name}"].values.astype(np.int32)
+                    )
+                    for structure_name in titles
                 }
+                # get granularity and n_sim_fit from the user
+                n_sim_fit, ok = QtWidgets.QInputDialog.getInt(
+                    self, "", "Number of simulations per tested combination",
+                    value=10, min=1, max=1000, step=1
+                )
+                if not ok:
+                    return
+                granularity, ok = QtWidgets.QInputDialog.getInt(
+                    self, "", "Granularity",
+                    value=21, min=1, max=1000, step=1
+                )
+                if not ok:
+                    return
+                self.n_sim_fit = n_sim_fit
+                self.granularity = granularity
                 # update the generate n structures button
                 n = len(self.N_structures_fit[self.structures[0].title])
                 estimated_time = self.estimate_fit_time(n)
@@ -3122,6 +3185,7 @@ class SimulationsTab(QtWidgets.QDialog):
             )
             if not save:
                 return
+            self.window.pwd = os.path.dirname(save)
 
         spinner = spinna.SPINNA(
             mixer=self.mixer,
@@ -3216,36 +3280,61 @@ class SimulationsTab(QtWidgets.QDialog):
 
     def save_fit_results(self) -> None:
         """Save fit results in .txt with all parameters used."""
+        metadata = self.summarize_fit_results()
+        out_path = self.structures_path.replace(".yaml", "_fit_summary.txt")
+        path, _ = QtWidgets.QFileDialog.getSaveFileName(
+            self, "Save fitting summary", out_path, filter="*.txt"
+        )
+        if path:
+            self.window.pwd = os.path.dirname(path)
+            with open(path, "w") as f:
+                for key, value in metadata.items():
+                    f.write(f"{key}: {value}\n")
+
+    def summarize_fit_results(self) -> None:
+        """Summarize fit results and parameters in a dictionary.
+
+        Returns
+        -------
+        metadata : dict
+            Dictionary with all parameters used and fit results.
+        """
         metadata = {}
         metadata["Date"] = datetime.now().strftime("%d/%m/%Y %H:%M:%S")
         metadata["File location of structures"] = self.structures_path
         metadata["Molecular targets"] = ", ".join(self.targets)
-        metadata["File location of experimental data"] = (
-            ", ".join([self.exp_data_paths[target] for target in self.targets])
-        )
-        if self.mask_den_stack.currentIndex() == 0:
-            metadata["File location of masks"] = (
-                ", ".join([self.mask_paths[target] for target in self.targets])
-            )
-            metadata["Number of simulations"] = self.n_sim_fit
-        else:
-            metadata["Simulated FOV (um)"] = (
-                ", ".join(
-                    [str(_/1e3) for _ in self.mixer.roi if _ is not None]
-                )
-            )
-        metadata["Label uncertainties (nm)"] = (
-            ", ".join([str(_.value()) for _ in self.label_unc_spins])
-        )
-        metadata["labeling efficiencies (%)"] = (
-            ", ".join([str(_.value()) for _ in self.le_spins])
-        )
+        metadata["Number of simulations"] = self.n_sim_fit
+        metadata["Parameter search space granularity"] = self.granularity
+        metadata["Dimensionality"] = self.dim_widget.currentText()
         metadata["Rotations mode"] = (
             self.settings_dialog.rot_dim_widget.currentText()
         )
-        metadata["Dimensionality"] = self.dim_widget.currentText()
-        metadata["Number of simulation repeats"] = str(self.n_sim_fit)
-        metadata["Parameter search space granularity"] = self.granularity
+        for t_idx, target in enumerate(self.targets):
+            metadata[f"File location of experimental data ({target})"] = (
+                self.exp_data_paths[target]
+            )
+            metadata[f"Label uncertainties (nm) ({target})"] = (
+                self.label_unc_spins[t_idx].value()
+            )
+            metadata[f"Labeling efficiencies (%) ({target})"] = (
+                self.le_spins[t_idx].value()
+            )
+        # masked used / observed density
+        if self.mask_den_stack.currentIndex() == 0:  # mask
+            for target in self.targets:
+                metadata[f"Mask ({target})"] = self.mask_paths[target]
+        else:
+            for target, den_spin in zip(self.targets, self.densities_spins):
+                if self.dim_widget.currentIndex() == 0:  # 2D
+                    label = f"Observed density (um^-2) ({target})"
+                if self.dim_widget.currentIndex() == 1:  # 3D
+                    label = f"Observed density (um^-3) ({target})"
+                metadata[label] = den_spin.value()
+            metadata["Simulated FOV (um)"] = " x ".join([
+                        str(_ / 1e3)
+                        for _ in self.mixer.roi if _ is not None
+                    ])
+        # fit results
         metadata["Best fitting proportions (%)"] = (
             self.fit_results_display.text().replace("\n", "")
         )
@@ -3285,25 +3374,18 @@ class SimulationsTab(QtWidgets.QDialog):
         for i, t1 in enumerate(self.mixer.targets):
             for t2 in self.mixer.targets[i:]:
                 key = f"{t1}-{t2}"
-                metadata[f"Number of neighbors at fitting ({key})"] = (
+                metadata[f"Number of neighbors considered ({key})"] = (
                     self.mixer.get_neighbor_counts(t1, t2)
                 )
 
         # labeling efficiency fitting
         if self.le_fitting_check.isChecked():
-            metadata["Labeling efficiency fitting"] = (
+            for target in self.targets:
+                metadata.pop(f"Labeling efficiency (%) ({target})", None)
+            metadata["Best fitting labeling efficiencies (%)"] = (
                 self.fit_results_display.text()
             )
-
-        # save metadata
-        out_path = self.structures_path.replace(".yaml", "_fit_summary.txt")
-        path, _ = QtWidgets.QFileDialog.getSaveFileName(
-            self, "Save fitting summary", out_path, filter="*.txt"
-        )
-        if path:
-            with open(path, "w") as f:
-                for key, value in metadata.items():
-                    f.write(f"{key}: {value}\n")
+        return metadata
 
     @check_structures_loaded
     @check_exp_data_loaded
@@ -3463,6 +3545,7 @@ class SimulationsTab(QtWidgets.QDialog):
             )
             if not path:
                 return
+            self.window.pwd = os.path.dirname(path)
         else:
             path = ""
 
@@ -3757,6 +3840,18 @@ class SimulationsTab(QtWidgets.QDialog):
         ok = True if isclose(sum_, 100, abs_tol=1e-3) else False
         return ok, sum_
 
+    def check_prop_str_sum(self, value: float, name: str) -> None:
+        """Stop the user from increasing the proportions to exceed
+        100%."""
+        sum_ = sum([_.value() for _ in self.prop_str_input_spins])
+        if sum_ <= 100:
+            return
+        all_names = [_.title for _ in self.structures]
+        idx = all_names.index(name)
+        self.prop_str_input_spins[idx].setValue(
+            self.prop_str_input_spins[idx].value() - (sum_ - 100)
+        )
+
     def find_roi(
         self,
         mode: Literal['fit', 'single_sim'] = 'fit',
@@ -3908,9 +4003,18 @@ class SimulationsTab(QtWidgets.QDialog):
             )
             if not path:
                 return
+            self.window.pwd = os.path.dirname(path)
         else:
             return
 
+        if self.mixer is None:
+            # set up a dummy mixer to get neighbor idx
+            self.mixer = spinna.StructureMixer(
+                structures=self.structures,
+                label_unc={"ALL": 0},
+                le={"ALL": 1.0},
+                width=10, height=10, depth=10,
+            )
         i = 0
         for (t1, t2, _) in self.mixer.get_neighbor_idx(duplicate=True):
             fig = self.nnd_plots[i]
@@ -3928,8 +4032,18 @@ class SimulationsTab(QtWidgets.QDialog):
             )
             if not path:
                 return
+            self.window.pwd = os.path.dirname(path)
         else:
             return
+
+        # set up a dummy mixer to get neighbor idx
+        if self.mixer is None:
+            self.mixer = spinna.StructureMixer(
+                structures=self.structures,
+                label_unc={"ALL": 0},
+                le={"ALL": 1.0},
+                width=10, height=10, depth=10,
+            )
 
         i = 0
         for (t1, t2, n) in self.mixer.get_neighbor_idx(duplicate=True):
@@ -3990,6 +4104,9 @@ class Window(QtWidgets.QMainWindow):
         self.resize(1024, 768)
         self.setMinimumSize(1024, 768)
         self.setMaximumSize(1024, 768)
+        settings = io.load_user_settings()
+        spinna_settings = settings.get("SPINNA", {})
+        self.pwd = spinna_settings.get("PWD", os.getcwd())
 
         # TABS
         self.tabs = QtWidgets.QTabWidget()
@@ -4029,6 +4146,14 @@ class Window(QtWidgets.QMainWindow):
 
         menu_bar = self.menuBar()
         self.plugin_menu = menu_bar.addMenu("Plugins")  # do not delete
+
+    def closeEvent(self, event: QtGui.QCloseEvent) -> None:
+        """Save the most recent directory in the user settings on
+        close."""
+        settings = io.load_user_settings()
+        settings["SPINNA"]["PWD"] = self.pwd
+        io.save_user_settings(settings)
+        QtWidgets.qApp.closeAllWindows()
 
 
 def main():